move population glm tests to a different file, adding TestGLMShared test class
